Changes for Adding entity classes

+++ src/main/sql/group_contents.sql
@@ -1,7 +1,7 @@
 CREATE TABLE group_contents (
   content_id int NOT NULL AUTO_INCREMENT,
   group_id int NOT NULL,
-  user_id int NOT NULL,
+  user_id BIGINT(20) NOT NULL,
   content_type varchar(50) NOT NULL,
   content varchar(500) not NULL,
   description varchar(500) not NULL,
